requirement.txt
streamlit
google-generativeai
python-dotenv
pyodbc